Switched Carousel and fixed products page UI
Previous Carousel had a bug in which sliding effect played to the wrong side when there are more then 1 item on a slide.

Changed the Product page view so that it looks better on tablet and mobile.
Made it so that img always takes up whole container, without being squashed and is centered instead.
Arrows were not working properly so I fixed it as well.
Deleted the use-fade hook because the animation did neither look good nor match the img loading time.
Img also scales with screen's width.