Deleted ProductStats
As this page was not in th einitial desig, but had to do some for of CRUD for project we decided to make it.
However, now it requires way to much work for little results.

As it is not mandatory for this shopping website to work I decided to delete it.